Final Solution
Learned how/when to use Hashset.  Getting a bit better at Linked Lists.
 Previous solution still works, but it’s pretty convoluted method.

The previous method was:
1) read into Hashtable, assign a value of # of instances it repeated.
2) take out only the keys that had values where the # of instances was
> 1
3) Return those keys as an arrayList
4) Run those arrayList against the original linked list, then for every
one time it spots one of those integers, it skips to the next one
5) However, if the next one it’s supposed to skill to is also a
duplicate, then it goes to the one after.

Super convoluted. Took a long time.  Good enough to be a p-set.