Find Roster Intersection & Assign Rank to Players
Out of two rosters Strings, find the players that intersect between the two and return them.  Then, take in a pipe-delimitated array and then assign a rank to each of the players based on their score. If the score is the same, then they are given the same rank and you skip +1 rank unit you reach the next one, therefore 70, 70, both players receive a rank of 1, but then 82 (3rd one) receives a rank of 3.